requirements.txt
discord.py==1.7.3
pypi-json==0.2.0
python-dotenv==0.19.2
riotwatcher==3.2.0
datetime==4.3
asyncio==3.4.3
DiscordUtils==1.3.4
python-math==0.0.1
PyMySQL==1.0.2